LayoutTests/fast/dom/node-childNodes-idempotence.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This test verifies that Node.childNodes returns the same NodeList when called repeatedly.");

shouldBe("document.documentElement.childNodes", "document.documentElement.childNodes");
shouldBeTrue("document.documentElement.childNodes === document.documentElement.childNodes");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
