third_party/WebKit/LayoutTests/fast/dom/node-childNodes-idempotence-expected.txt

This test verifies that Node.childNodes returns the same NodeList when called repeatedly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.documentElement.childNodes is document.documentElement.childNodes
PASS document.documentElement.childNodes === document.documentElement.childNodes is true
PASS successfullyParsed is true

TEST COMPLETE
